[not verified] jetpack: qr-post: refactoring -> move helper function to utils
