+++ projects/plugins/jetpack/extensions/plugins/post-publish-qr-post-panel/components/image-actions-panel-row.js
@@ -7,37 +7,9 @@ import { useSelect } from '@wordpress/data';
 import { store as editorStore } from '@wordpress/editor';
 
 /**
- * Handler function to create an image from the QR code.
- *
- * @param {string} slug      - The slug of the image to create.
- * @param {object} ref       - The ref of the QR code DOM element.
- * @param {boolean} download - Whether to download the image. Defaults to true.
- * @returns {void}           - Nothing when the image is not created.
+ * Internal dependencies
  */
-function handleDownloadQRCode( slug, ref, download = true ) {
-	if ( ! slug ) {
-		return;
-	}
-
-	if ( ! ref?.current ) {
-		return;
-	}
-
-	const canvasElement = ref.current.querySelector( 'canvas' );
-	if ( ! canvasElement ) {
-		return;
-	}
-
-	// Convert to canvas element to data URL image.
-	canvasElement.toBlob( imageBlob => {
-		const imageURL = URL.createObjectURL( imageBlob );
-		const tempLink = document.createElement( 'a' );
-		tempLink.href = imageURL;
-		// Download, or not.
-		tempLink.setAttribute( download ? 'download' : 'target', `qr-post-${ slug }.png` );
-		tempLink.click();
-	} );
-}
+import handleDownloadQRCode from '../utils/handle-download-qrcode.js';
 
 /**
  * Row panel component to address the actions
